# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Automation Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [auth]
      summary: Create a LemonSqueezy checkout session to upgrade plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                planKey:
                  type: string
                  enum: [hobby, pro, team]
              required: [planKey]
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required: [url]
        "401":
          description: Unauthorized
  /billing/portal:
    post:
      operationId: createPortalSession
      tags: [auth]
      summary: Get LemonSqueezy customer portal URL
      responses:
        "200":
          description: Portal session URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required: [url]
        "401":
          description: Unauthorized
  /billing/usage:
    get:
      operationId: getBillingUsage
      tags: [auth]
      summary: Get current compute-time usage for the billing period
      responses:
        "200":
          description: Usage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingUsage"
        "401":
          description: Unauthorized
  /billing/products:
    get:
      operationId: listBillingProducts
      tags: [auth]
      summary: List Stripe products with prices
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BillingProduct"
                required: [data]
        "401":
          description: Unauthorized
...
  /executions:
    get:
      operationId: listExecutions
      tags: [executions]
      summary: List workflow executions
      parameters:
        - name: workflowId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [success, error, running, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Execution"
  /executions/{id}:
    get:
      operationId: getExecution
      tags: [executions]
      summary: Get execution details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
  /executions/{id}/retry:
    post:
      operationId: retryExecution
      tags: [executions]
      summary: Retry a failed execution with its original input data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New execution created for the retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "404":
          description: Execution not found
        "400":
          description: Execution is not in a retryable state
...
  /credentials:
    get:
      operationId: listCredentials
      tags: [credentials]
      summary: List stored credentials for the current user
      responses:
        "200":
          description: List of credentials (data field is redacted)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credential"
        "401":
          description: Unauthorized
    post:
      operationId: createCredential
      tags: [credentials]
      summary: Store a new credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialBody"
      responses:
        "201":
          description: Created credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "401":
          description: Unauthorized
  /credentials/{id}:
    delete:
      operationId: deleteCredential
      tags: [credentials]
      summary: Delete a stored credential
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List approval requests for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Approval requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalRequest"
  /approvals/{id}/respond:
    post:
      operationId: respondToApproval
      tags: [approvals]
      summary: Approve or reject a pending approval request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalResponse"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  decision:
                    type: string
                required: [success, decision]
        "404":
          description: Not found
        "409":
          description: Already responded
components:
  schemas:
...
    Execution:
      type: object
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        workflowName:
          type: string
        status:
          type: string
          enum: [success, error, running]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        nodeResults:
          type: array
          items:
            $ref: "#/components/schemas/NodeResult"
        inputData:
          type: object
        outputData:
          type: object
          nullable: true
      required: [id, workflowId, workflowName, status, startedAt, nodeResults, inputData]
    NodeResult:
      type: object
      properties:
        nodeId:
          type: string
        nodeType:
          type: string
        status:
          type: string
          enum: [success, error, skipped]
        output:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
        durationMs:
          type: integer
      required: [nodeId, nodeType, status, durationMs]
...
    BillingProduct:
      type: object
      properties:
        product_id:
          type: string
        product_name:
          type: string
        product_description:
          type: string
          nullable: true
        price_id:
          type: string
        unit_amount:
          type: integer
        currency:
          type: string
        recurring:
          type: object
          nullable: true
      required: [product_id, product_name, price_id, unit_amount, currency]
    BillingUsage:
      type: object
      properties:
        plan:
          type: string
          enum: [free, hobby, pro, team]
        planLabel:
          type: string
        computeSecondsUsed:
          type: integer
        computeSecondsLimit:
          type: integer
          nullable: true
          description: "null means unlimited"
        computeMsUsed:
          type: integer
        percentUsed:
          type: integer
        runsThisMonth:
          type: integer
        resetAt:
          type: string
          format: date-time
          nullable: true
        price:
          type: integer
      required: [plan, planLabel, computeSecondsUsed, computeMsUsed, percentUsed, runsThisMonth, price]
...
    Credential:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        credentialType:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, credentialType, createdAt]
    CreateCredentialBody:
      type: object
      properties:
        name:
          type: string
        credentialType:
          type: string
        data:
          type: object
          description: Credential data e.g. apiKey for Resend, url for Slack webhook
      required: [name, credentialType, data]
    ApprovalRequest:
      type: object
      properties:
        id:
          type: integer
        executionId:
          type: integer
        nodeId:
          type: string
        workflowId:
          type: integer
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        approverEmails:
          type: array
          items:
            type: string
        deadlineAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        decision:
          type: string
          nullable: true
        deciderEmail:
          type: string
          nullable: true
        responseNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, executionId, nodeId, workflowId, userId, title, message, approverEmails, status, createdAt]
    ApprovalResponse:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        responseNote:
          type: string
        deciderEmail:
          type: string
      required: [decision]